dbutils/requirements.txt
nose
mysql-connector
